Update getting started eksctl.yaml (#563)
* Update getting started eksctl.yaml

Switching from self managed node group to managed node group. This could help customers in the future if they keep a cluster around a while and want to update the EKS version.

Also specified the EKS version because depending on which version of `eksctl` the user has it may deploy a different version. I would use 1.21 but the official `eksctl` with 1.21 support isn't out yet. 1.20 has been covered for a while and gets us a semi-new version of a cluster to avoid unknown issues.

* Update commands to avoid errors

Depending on the users shell having a comment in a command block can cause an error. My zsh errored and didn't run the commands. Removing the extra comments to avoid that issue.

* Add helm chart version

Will also need to be versioned with new releases.