Make it possible to set node naming convention
As of k8s 1.23, CCM supports both instance ID (resource name) and private dns (ip name) based naming.
It is not possible for karpenter to determine when which convention is used, and this may also depend on the installer used.
Therefore this is added as a CLI argument

Update pkg/utils/options/options.go

Co-authored-by: Ellis Tarn <ellistarn@gmail.com>