merge bin-packing and scheduling
This solves some issues with topologies and prepares for pod-affinity in
that a separate bin-packing pass would break apart pods that the scheduler
intended to be packed together.  By combining bin-packing with scheduling
the 'scheduling nodes' created correspond to actual K8s nodes that will be
created via the cloud provider, eliminating any errors in counting
hostname topologies.