Add ability to configure Subnets and Security Groups by ID
This provides a way to configure the VPC Subnets used by Karpenter for
users who have clusters in environments where they do not have access to
the VPC Subnet tags.

This can happen in environments which use a [Shared VPC
Architecture][1]. To provide a workaround for this type of architecture,
the magic key `aws-ids` is used to indicate that the user is specifying
the AWS IDs explicitly.

In the implementation, the magic key indicates that a different type of
`ec2.Filter` object should be used in the call to the
`ec2.DescribeSubnetsWithContext` method.

[1]: https://docs.aws.amazon.com/vpc/latest/userguide/vpc-sharing.html#vpc-share-limitations

An identical feature is added for Security Group discovery.

Signed-off-by: Ira W. Snyder <ira@descarteslabs.com>