fix: add initial delay to liveness and let CP know it's run by the webhook (#2032)
The webhook just needs the CP for Validate/Default, so let the CP know
this to allow it to skip some processing that isn't required if it
chooses to. Add an initial delay to our liveness probes to ensure that
any initial startup doesn't lead to restarts on the webhook/controller.