Update api_version in helm provider
The api_version in the 'exec' section of the 'helm' provider was obsolete. From the kubernetes documentation, it should be 'v1beta1' or 'v1'.
The v1alpĥa1 will return the error: "no kind "ExecCredential" is registered for version "client.authentication.k8s.io/v1alpha1" in scheme "pkg/runtime/scheme.go:100"
The aws cli creates a EKS cluster where the 'v1alpha1' was deprecated and will not deploy.