fix: wait for node deletion before continuing consolidation (#2373)
ix: wait for node deletion before continuing consolidation

Removes a check for un-initialized nodes that would block all
consolidation if a node fails to start and re-uses the nominated
node check from the emptiness controller to accomplish
a similar task.  Upon node deletion, wait for the node to be
deleted to pause consolidation which ensures we don't overshoot
consolidation due to a PDB throttling the deletion of a node.