Changes to make logging easier with karpenter
Currently if karpenter is running in fargate the only way to use 3rd party logging would be to send all the logs via fargate logging on the cluster to firehose and then to actual app.

With this PR it's now possible to override zap logger to send logs to stdout and also to a file. With addition of sideCar config as well we can now run logging sidecar which shares volumes with karpenter container and use e.g. fluentbit sidecar to forward the logs to our monitoring tool of choice.