feat: adds metrics for instance type vcpu and memory
This exposes 2 new metrics with details on the number of VCPUs and memory
for an instance type.

With these bits of information, you could start to calculate a per VCPU cost
and a per GiB cost, which can give you a rough estimate of the cost of running
a Pod on a specific node.