fix: Prevent Machine Link Controller workqueue clogging
Now, to prevent requeueing of requests in the machine link
controller we return `reconcile.Result{RequeueAfter: math.MaxInt64}`,
which results in an ever-growing queue and memory leakage, as
it only delayed the requeueing requests with an extremely long wait time.

This PR improves the handling of requests requeueing by returning
an empty `reconcile.Result{}`, which already has Requeue flag set
to false.