doc: Add FAQ section to answer logging questions
Problem

The only ways to know how to configure the logging in Karpenter are to dig through past issues where configuration is discussed or to be aware of the underlying logging library and how to use it already. This can be a a higher burden than necessary to make small modifications to the logs.

Solution

Add a section to the FAQ about logging. This section links out to where to update, what to update, and how to update, to make changes to the logging configuration.

Result

Reading the FAQ will answer logging questions and direct people to the appropriate resources without needing to have a deep understanding of the code, dig through issues, or have prior knowledge of the logging library.