Fix script if there are multiple sunbets
The output of the subcommand should not be wrapped in `$()`. This means that the output of the subcommand will be treated as multiple separate arguments. Therefore, if the subcommand outputs multiple subnet IDs separated by spaces, they will be treated as separate arguments passed to `--resources`. This resolves the previous "InvalidID" error if we have multiple subnets for eks nodegroup.