Add interruption handling for rebalance recommendations
Karpenter currently cordons and drains nodes when a spot termination
event is received through the configured SQS queue.  It also adds
the instance type into a temporary block list to avoid cycling the
same instance types.

These are all behaviours which would be beneficial during a rebalance
recommendation event with the added benefit that rebalance
recommendations allow more time to drain the node.

It has been requested by members of the community that this feature
should be optional.  It makes sense to enable or disable it on a
per NodePool basis but no spec has been agreed yet.  This PR contains
code based on a possible API change.  Further discussion will be
required to agree the change and get the new API published.