fix: discover deprecated EKS optimized AMIs
In the event of an AMI deprecation, we may not discover the new SSM
parameter for up to 24 hours. Rather than failing to provision for
those 24 hours, we'll continue to discover and use the deprecated AMI.
This only impacts those using Karpenter's automatic AMI upgrade system
which is not meant for production workloads.