Fix EC2NodeClass readiness to not stuck in False
We observed some EC2NodeClasses stuck in Ready=False with message of `Failed to detect the cluster CIDR`. We believe this is caused by a bug in Karpenter, and this commit fixes that.

The previous commit adds a failing test to reproduce and the demonstrate the issue, and this commit fixes the implementation.